+++ app/models/pay/subscription.rb
@@ -4,6 +4,7 @@ module Pay
 
     # Associations
     belongs_to :customer
+    belongs_to :payment_method, optional: true, primary_key: :processor_id
     has_many :charges
 
     # Scopes
@@ -28,7 +29,6 @@ module Pay
 
     store_accessor :data, :paddle_update_url
     store_accessor :data, :paddle_cancel_url
-    store_accessor :data, :stripe_account
     store_accessor :data, :subscription_items
 
     attribute :prorate, :boolean, default: true